# backend/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "API для управления складом товаров с ролевой моделью доступа",
    "title": "MarketEase API",
    "contact": {},
    "version": "2.0"
  },
  "host": "localhost:8080",
  "basePath": "/api",
  "paths": {
...
    "/profile": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Получить информацию о своём профиле",
        "produces": ["application/json"],
        "tags": ["profile"],
        "summary": "Профиль пользователя",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/models.User"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          }
        }
      }
    },
    "/profile/password": {
      "put": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Изменить свой пароль",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["profile"],
        "summary": "Смена пароля",
        "parameters": [
          {
            "description": "Старый и новый пароль",
            "name": "input",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/controllers.UpdatePasswordInput"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "controllers.UpdatePasswordInput": {
      "type": "object",
      "required": ["newPassword", "oldPassword"],
      "properties": {
        "newPassword": {
          "type": "string",
          "minLength": 6,
          "example": "newpassword456"
        },
        "oldPassword": {
          "type": "string",
          "example": "oldpassword123"
        }
      }
    },
...
    "models.ErrorResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "Произошла ошибка"
        }
      }
    },
...
    "models.User": {
      "type": "object",
      "properties": {
        "createdAt": {
          "type": "string"
        },
        "deletedAt": {
          "type": "string"
        },
        "email": {
          "type": "string",
          "example": "john@example.com"
        },
        "id": {
          "type": "string",
          "example": "e7b1c2f1-567b-4b14-a8d7-cd5b08bfc9d0"
        },
        "role": {
          "type": "string",
          "example": "viewer"
        },
        "status": {
          "type": "string",
          "example": "active"
        },
        "updatedAt": {
          "type": "string"
        },
        "username": {
          "type": "string",
          "example": "john_doe"
        }
      }
    }
  },
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}